+++ application/views/fleetdetail.php
@@ -1,0 +1,39 @@
+<div class="row">
+
+    <div class="card-body">
+        <div class="table-responsive">
+            <table class="table table-default" id="dataTable" width="100%" cellspacing="0">
+                    <thead>
+                    <tr>
+                        <th>ID</th>
+                        <th>Plane Name</th>
+                        <th>Manufacturer</th>
+                        <th>Model</th>
+                        <th>Price</th>
+                        <th>Seats</th>
+                        <th>Reach</th>
+                        <th>Cruise</th>
+                        <th>Takeoff</th>
+                        <th>Hourly</th>
+                    </tr>
+                    </thead>
+                    <tbody>
+                    <th>{id}</th>
+                    <th>{planeName}</th>
+                    <th>{manufacturer}</th>
+                    <th>{model}</th>
+                    <th>{price}</th>
+                    <th>{seats}</th>
+                    <th>{reach}</th>
+                    <th>{cruise}</th>
+                    <th>{takeoff}</th>
+                    <th>{hourly}</th>
+                    </tbody>
+            <table/>
+                <div class="text-center">
+                <a href="/fleet">Go back</a>
+                </div>
+        </div>
+
+    </div>
+</div>